pyproject.toml
[tool.poetry]
name = "jacobson"
version = "0.1.0"
description = "self-hosted zipcode API"
authors = ["Christian G. Semke <chrisemke@protonmail.com>"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
fastapi-slim = "^0.111.0"
strawberry-graphql = "^0.232.1"
httpx = "^0.27.0"
pydantic = "^2.7.1"
alembic = "^1.13.1"
asyncmy = "^0.2.9"
pydantic-settings = "^2.2.1"
sqlmodel-slim = {extras = ["standard"], version = "^0.0.18"}
sqlmodel = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.5"
mypy = "^1.10.0"
commitizen = "^3.27.0"
poetry-plugin-export = "^1.8.0"
fastapi = "^0.111.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.25"
mkdocstrings = "^0.25.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.7"
respx = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
exclude = ["build", "docker", "tests"]
plugins = ["pydantic.mypy", "strawberry.ext.mypy_plugin"]

[tool.ruff]
lint.select = [
    "E",   # pydocstyle
    "W",   # pydocstyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "D",   # docstrings
    "RUF", # ruff
]
lint.ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D211", # no-blank-line-before-class
    "D212", # Multi-line docstring summary should start at the first line
]
line-length = 79
target-version = "py312"

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
minversion = "8.2"
addopts = "-vv --no-header --ignore=docker"
testpaths = ["tests"]
